# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jargon GTM Execution API",
    "version": "1.0.0",
    "description": "Public API for deploying workspaces and executing outbound (queue, email, voice, disposition). Tenancy is derived from the API key. Do not send org_id."
  },
  "servers": [
    {
      "url": "https://jargon-api-production.up.railway.app/v1",
      "description": "Production"
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/suppressions": {
      "get": {
        "operationId": "listSuppressions",
        "tags": ["Compliance"],
        "summary": "This org's do-not-contact list. Global hard bounces apply but are not listed.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["suppressions"],
                  "properties": {
                    "suppressions": { "type": "array", "items": { "$ref": "#/components/schemas/Suppression" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "addSuppressions",
        "tags": ["Compliance"],
        "summary": "Block emails, domains, phone numbers (do-not-call), or LinkedIn URLs. Cancels queued messages they cover.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddSuppressionsRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["added", "invalid"],
                  "properties": {
                    "added": { "type": "array", "items": { "$ref": "#/components/schemas/Suppression" } },
                    "invalid": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/suppressions/{id}": {
      "delete": {
        "operationId": "removeSuppression",
        "tags": ["Compliance"],
        "summary": "Remove an org suppression",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "204": { "description": "Removed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "jarg_",
        "description": "API key from the dashboard or `jargon api-keys create`. Tenancy is implied."
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or non-API-key bearer token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NotFound": {
        "description": "Not found in this org",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Error": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "billingUrl": { "type": "string" },
          "remaining": { "type": "number" }
        }
      },
...
      "Suppression": {
        "type": "object",
        "required": ["id", "orgId", "kind", "value", "reason", "createdAt"],
        "properties": {
          "id": { "type": "string" },
          "orgId": { "type": ["string", "null"] },
          "kind": { "type": "string", "enum": ["email", "domain", "phone", "linkedin"] },
          "value": { "type": "string" },
          "reason": { "type": "string", "enum": ["unsubscribed", "bounced", "complaint", "do_not_call", "manual"] },
          "source": { "type": "string" },
          "createdAt": { "type": "integer" }
        }
      },
      "AddSuppressionsRequest": {
        "type": "object",
        "required": ["kind"],
        "additionalProperties": false,
        "properties": {
          "kind": { "type": "string", "enum": ["email", "domain", "phone", "linkedin"] },
          "value": { "type": "string" },
          "values": { "type": "array", "maxItems": 5000, "items": { "type": "string" } },
          "reason": {
            "type": "string",
            "enum": ["unsubscribed", "do_not_call", "manual"],
            "description": "Defaults to do_not_call for phone, manual otherwise"
          }
        }
      },
...
    }
  }
}